feat: tested everything works, added security to cruds

+++ src/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
 
@@ -7,29 +7,78 @@ export class UsersService {
   constructor(private prisma: PrismaService) {}
 
   async getAllUsers() {
-    return this.prisma.users.findMany({
+    const users = await this.prisma.users.findMany({
       orderBy: {
         created_at: 'desc',
       },
     });
+
+    if (users.length === 0) {
+      throw new NotFoundException(
+        "Aucun utilisateurs existants ! Veuillez contacter l'administrateur car cette erreur devrait être impossible !",
+      );
+    }
+    return {
+      status: 'Succès',
+      data: users,
+    };
   }
 
   async getOneUser(id: string) {
-    return this.prisma.users.findUnique({
+    const existingUser = await this.prisma.users.findUnique({
       where: { id_user: id },
     });
+
+    if (!existingUser) {
+      throw new NotFoundException('Utilisateur inexistant !');
+    }
+
+    const user = await this.prisma.users.findUnique({
+      where: { id_user: id },
+    });
+
+    return {
+      status: 'Succès',
+      data: user,
+    };
   }
 
   async updateUser(id: string, dto: UpdateUserDto) {
-    return this.prisma.users.update({
+    const existingUser = await this.prisma.users.findUnique({
+      where: { id_user: id },
+    });
+
+    if (!existingUser) {
+      throw new NotFoundException('Utilisateur inexistant !');
+    }
+
+    await this.prisma.users.update({
       where: { id_user: id },
       data: { ...dto },
     });
+
+    return {
+      status: 'Succès',
+      message: 'Utilisateur modifié !',
+    };
   }
 
   async deleteUser(id: string) {
-    return this.prisma.users.delete({
+    const existingUser = await this.prisma.users.findUnique({
       where: { id_user: id },
     });
+
+    if (!existingUser) {
+      throw new NotFoundException('Utilisateur inexistant !');
+    }
+
+    await this.prisma.users.delete({
+      where: { id_user: id },
+    });
+
+    return {
+      status: 'Succès',
+      message: 'Utilisateur supprimé !',
+    };
   }
 }
